% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jqinit}
\alias{jqinit}
\title{Initializes the JQ system
the handle is stored under the name \code{jqhandle} in options}
\usage{
jqinit()
}
\value{
the handle itself
}
\description{
Initializes the JQ system
the handle is stored under the name \code{jqhandle} in options
}

