% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jqparse}
\alias{jqparse}
\title{Will apply a query to a JSON object and return a json}
\usage{
jqparse(f, query, jq = options("jqhandle")[[1]])
}
\arguments{
\item{f}{is a JSON string or NULL (see details)}

\item{query}{is either a string (which will be compiled via \code{jqquery}) or a an object returned by \code{jqquery}}

\item{jq}{is the handle}
}
\value{
a json string with the results of the query
}
\description{
Will apply a query to a JSON object and return a json
}
\details{
In a RHIPE mapreduce job, you can leave \code{f} null,
provided you called \code{rhStringMapValues} before. The function \code{rhStringMapValues} will create a big json array out of \code{map.values} and store it internally for JQ to process. If \code{f} is NULL, JQ will apply the \code{query} to this internal JSON array version of \code{map.values}
}
\seealso{
\code{\link{jqquery}}
}

