% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jqpp}
\alias{jqpp}
\title{pretty print a JSON}
\usage{
jqpp(s, print = TRUE, color = TRUE, ordered = TRUE,
  jq = options("jqhandle")[[1]])
}
\arguments{
\item{s}{the json}

\item{print}{to print ot not?}

\item{color}{use color text (wont work in Rstudio), terminal only.}

\item{ordered}{sort the keys?}
}
\value{
the prettified json
}
\description{
pretty print a JSON
}

