% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jqquery}
\alias{jqquery}
\title{Stores a query in the jq system, so that it doesn't have to be compiled repeatedly}
\usage{
jqquery(f, jq = options("jqhandle")[[1]])
}
\arguments{
\item{f}{is the query string}

\item{jq}{is the handle}
}
\value{
an object you pass to \code{jqparse}
}
\description{
Stores a query in the jq system, so that it doesn't have to be compiled repeatedly
}
\details{
see \href{http://stedolan.github.io/jq/tutorial/} for details of the query
}
\seealso{
\code{\link{jqparse}}
}

