% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhStringMapValues}
\alias{rhStringMapValues}
\title{makes a json array version of a map.values (of FHR objects)}
\usage{
rhStringMapValues(f, jq = options("jqhandle")[[1]])
}
\arguments{
\item{f}{is a list of JSON strings (e.g. map.values when reading from FHR data)}
}
\description{
makes a json array version of a map.values (of FHR objects)
}
\details{
use this if you'd like to apply a JQ query to the JSON array of map.values FHR JSON objects. If you do, call \code{jqparse} with it's first argument as NULL.
}
\seealso{
\code{\link{jqquery}},\code{\link{jqparse}}
}

